requirements.txt
Pillow==11.3.0
tkinterdnd2==0.4.3
